# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinePilot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  avatarUrl?: string | null;
  provider?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ServerStatus = (typeof ServerStatus)[keyof typeof ServerStatus];

export const ServerStatus = {
  online: "online",
  offline: "offline",
  connecting: "connecting",
  error: "error",
} as const;

export interface Server {
  id: number;
  userId: number;
  name: string;
  host: string;
  sshPort: number;
  sshUsername: string;
  status: ServerStatus;
  serverType?: string | null;
  gameMode?: string | null;
  createdAt: string;
}

export interface CreateServerRequest {
  name: string;
  host: string;
  sshPort: number;
  sshUsername: string;
  sshPassword: string;
  serverType?: string | null;
  gameMode?: string | null;
}

export interface UpdateServerRequest {
  name?: string;
  host?: string;
  sshPort?: number;
  sshUsername?: string;
  sshPassword?: string;
  serverType?: string | null;
  gameMode?: string | null;
}

export interface ConnectionStatus {
  connected: boolean;
  message: string;
}

export interface ServerStats {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  ramUsed: number;
  ramTotal: number;
  cpuUsage: number;
  tps: number;
  diskUsed: number;
  diskTotal: number;
}

export type SetupServerRequestServerType =
  (typeof SetupServerRequestServerType)[keyof typeof SetupServerRequestServerType];

export const SetupServerRequestServerType = {
  paper: "paper",
  spigot: "spigot",
  purpur: "purpur",
} as const;

export type SetupServerRequestGameMode =
  (typeof SetupServerRequestGameMode)[keyof typeof SetupServerRequestGameMode];

export const SetupServerRequestGameMode = {
  survival: "survival",
  skyblock: "skyblock",
  smp: "smp",
} as const;

export interface SetupServerRequest {
  serverType: SetupServerRequestServerType;
  gameMode: SetupServerRequestGameMode;
  installPlugins?: string[];
}

export interface Plugin {
  id: string;
  name: string;
  description: string;
  version: string;
  author: string;
  category: string;
  downloadUrl: string;
  tags: string[];
}

export interface InstalledPlugin {
  id: number;
  serverId: number;
  pluginId: string;
  pluginName: string;
  version: string;
  installedAt: string;
}

export interface InstallPluginRequest {
  serverId?: number | null;
  pluginId: string;
}

export interface ConsoleCommandRequest {
  serverId: number;
  command: string;
}

export interface ConsoleCommandResponse {
  output: string;
  success: boolean;
}

export interface ConsoleEntry {
  id: number;
  serverId: number;
  command: string;
  output: string;
  createdAt: string;
}

export interface FileEntry {
  name: string;
  path: string;
  isDirectory: boolean;
  size?: number | null;
  modifiedAt?: string | null;
}

export interface FileContent {
  path: string;
  content: string;
}

export interface WriteFileRequest {
  serverId: number;
  path: string;
  content: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatRequest {
  messages: ChatMessage[];
  serverId?: number | null;
}

export interface AnalyzeLogRequest {
  log: string;
}

export interface SuggestConfigRequest {
  pluginName: string;
  serverId?: number | null;
}

export interface AiChatResponse {
  message: string;
}

export interface ActivityLog {
  id: number;
  userId: number;
  serverId?: number | null;
  action: string;
  details?: string | null;
  createdAt: string;
}

export type GoogleCallbackParams = {
  code?: string;
  state?: string;
};

export type GithubCallbackParams = {
  code?: string;
  state?: string;
};

export type ListInstalledPluginsParams = {
  serverId?: number | null;
};

export type GetConsoleHistoryParams = {
  serverId: number;
};

export type ListFilesParams = {
  serverId: number;
  path?: string;
};

export type ReadFileParams = {
  serverId: number;
  path: string;
};

export type DeleteFileParams = {
  serverId: number;
  path: string;
};

export type ListActivityParams = {
  serverId?: number;
  limit?: number;
};