# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MinePilot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/delete-account:
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Delete account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /auth/google/callback:
    get:
      operationId: googleCallback
      tags: [auth]
      summary: Google OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect with token
...
  /auth/github/callback:
    get:
      operationId: githubCallback
      tags: [auth]
      summary: GitHub OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect with token
...
  /console/command:
    post:
      operationId: sendCommand
      tags: [console]
      summary: Send command to server console
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsoleCommandRequest"
      responses:
        "200":
          description: Command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleCommandResponse"
  /console/history:
    get:
      operationId: getConsoleHistory
      tags: [console]
      summary: Get console command history
      security:
        - bearerAuth: []
      parameters:
        - name: serverId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Command history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsoleEntry"
  /files/list:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in directory
      security:
        - bearerAuth: []
      parameters:
        - name: serverId
          in: query
          required: true
          schema:
            type: integer
        - name: path
          in: query
          schema:
            type: string
            default: "/"
      responses:
        "200":
          description: File listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileEntry"
  /files/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read file content
      security:
        - bearerAuth: []
      parameters:
        - name: serverId
          in: query
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write file content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteFileRequest"
      responses:
        "200":
          description: File written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /files/delete:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      security:
        - bearerAuth: []
      parameters:
        - name: serverId
          in: query
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, createdAt]
...
    ConsoleCommandRequest:
      type: object
      properties:
        serverId:
          type: integer
        command:
          type: string
      required: [serverId, command]
    ConsoleCommandResponse:
      type: object
      properties:
        output:
          type: string
        success:
          type: boolean
      required: [output, success]
    ConsoleEntry:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        command:
          type: string
        output:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, serverId, command, output, createdAt]
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        isDirectory:
          type: boolean
        size:
          type: integer
          nullable: true
        modifiedAt:
          type: string
          nullable: true
      required: [name, path, isDirectory]
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required: [path, content]
    WriteFileRequest:
      type: object
      properties:
        serverId:
          type: integer
        path:
          type: string
        content:
          type: string
      required: [serverId, path, content]
...